% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_GeneticsParams.R
\docType{class}
\name{CorrRWTraits}
\alias{CorrRWTraits}
\title{Set genetic traits structure for CorrRW traits}
\usage{
CorrRWTraits(Positions = list("random","random"), NbOfPositions = c(10, 10),
ExpressionType = rep("additive",2),
InitialDistribution = rep("normal",2), InitialParameters = matrix(c(rep(0.5,2),(rep(0.1,2), nrow=2),
IsInherited = rep(FALSE,2),
MutationDistribution = rep("normal",2), MutationParameters = matrix(c(rep(0.5,2),(rep(0.2,2), nrow=2),
MutationRate = rep(0.0001,2), OutputValues = rep(FALSE,2))
}
\arguments{
\item{Positions}{Loci positions coding for the trait within genome. Should be provided as a list. Entries can either be a string (\code{"random"}) and/or vectors of integers.
The length must be 2 for the required CorrRW traits \code{Steplength} and \code{Rho}.}

\item{NbOfPositions}{Only specify when the \code{Positions} of the CorrRW trait is set to ‘random’, else must be blank (NULL).
The length must be 2 for the required CorrRW traits \code{Steplength} and \code{Rho}.}

\item{ExpressionType}{Type of expression for the emigration trait. Can be either \code{additive} or \code{average}.
The length must be 2 for the required CorrRW traits \code{Steplength} and \code{Rho}.}

\item{InitialDistribution}{Distribution of the initial values. Can be \code{uniform} or \code{normal}. Should be provided as a vector of strings.
The length must be 2 for the required CorrRW traits \code{Steplength} and \code{Rho}.}

\item{InitialParameters}{Parameters for the initial distribution: You must provide two colums min and max  for \code{uniform} distribution and mean and sd for \code{normal} distribution.
Each row in the matrix corresponds to an CorrRW trait. The number of rows must be 2 for the required CorrRW traits \code{Steplength} and \code{Rho}.}

\item{IsInherited}{Should the emigration trait be inherited? Can be either \code{TRUE} or \code{FALSE}.
The length must be 2 for the required CorrRW traits \code{Steplength} and \code{Rho}.}

\item{MutationDistribution}{Distribution for mutations to draw from. Can be \code{uniform} or \code{normal}.
The length must be 2 for the required CorrRW traits \code{Steplength} and \code{Rho}.}

\item{MutationParameters}{Parameters for the mutation distribution: You must provide two colums: min and max for \code{uniform} distribution and mean and sd for \code{normal} distribution.
Each row in the matrix corresponds to an CorrRW trait. The number of rows must be 2 for the required CorrRW traits \code{Steplength} and \code{Rho}.}

\item{MutationRate}{Mutation rate applicable to this type of loci. Must be between 0.0 and 1.0.
The length must be 2 for the required CorrRW traits \code{Steplength} and \code{Rho}.}

\item{OutputValues}{If OutputGeneValues in GeneticsFile is enabled, should allele values for this gene be written to output? Ignored if OutputGeneValues is set to \code{FALSE}.
The length must be 2 for the required CorrRW traits \code{Steplength} and \code{Rho}.}
}
\value{
a parameter object of class "CorrRWTraitsParams"
}
\description{
Depending on the settings of \code{\link[RangeShiftR]{CorrRW}}, the following traits can be evolvable:

- The length of a single transfer step \code{Steplength} \cr
- The angle correlation parameter of the random walk \code{Rho} \cr

The entries of the trait parameters must be provided in the same order as the CorrRW traits are listed above (first \code{StepLength} and second \code{Rho}). If parameters expect a matrix, the rows must match the order of CorrRW traits listed above.
}
\details{
Traits set to evolve cannot simultaneously be stage-dependent.

The alleles of each trait can be expressed according to an additive model (allele values across all loci are summed) or be averaged.

Mutations are additive and can be sampled in either a \code{normal} or \code{uniform} distribution.

Initial allele values are sampled in a \code{normal} or \code{uniform} distribution.

CorrRW traits can also be **not** inherited, that is, allele values are resampled from the initial distribution for every new individual.

Dominance values are not applicable for CorrRW traits.
}
\author{
Jette Reeg
}
