% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_handling.R
\name{createODD}
\alias{createODD}
\title{Creating ODD template file for a specific RangeShiftR parameter master object}
\usage{
createODD(filename, s, type, title)
}
\arguments{
\item{filename}{Name of the R markdown file and document to be created, must have ending ".Rmd", e.g. 'ODD_protocol.Rmd'}

\item{s}{RangeShiftR parameter object}

\item{type}{file type of the rendering process output. Can currently only be "pdf_document"}

\item{title}{Title of the study if possible including authors, e.g. "'Study title ABC' by author ABC}
}
\description{
This function creates an ODD template file for a specific RangeShiftR parameter master object \code{s}.
It only creates a new file, if the \code{filename} doesn't exist in the folder.
If the \code{filename} already exists, it only renders the document to the given type.
}
