% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_GeneticsParams.R
\docType{class}
\name{CRWTraits}
\alias{CRWTraits}
\title{Set genetic traits structure for CRW traits}
\value{
a parameter object of class "CRWTraitsParams"
}
\description{
Can I have a list of different CRWtraits() depending on the dispersal parameter? (E0. Ealpa, Ebeta,)
}
\author{
Jette Reeg
}
