% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_GeneticsParams.R
\docType{class}
\name{EmigrationTraits}
\alias{EmigrationTraits}
\title{Set genetic traits structure for Emigration traits}
\value{
a parameter object of class "EmigrationTraitsParams"
}
\description{
Set genetic traits structure for Emigration traits
}
\author{
Jette Reeg
}
