% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_GeneticsParams.R
\docType{class}
\name{GeneticLoadTraits}
\alias{GeneticLoadTraits}
\title{Set genetic traits structure for genetic fitness}
\value{
a parameter object of class "GeneticLoadParams"
}
\description{
Set genetic traits structure for genetic fitness
}
\author{
Jette Reeg
}
