% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_handling.R
\name{getLocalisedEquilPop}
\alias{getLocalisedEquilPop}
\title{Calculates the equilibrium distribution for a localised (i.e. non-spatial) closed population}
\usage{
getLocalisedEquilPop(demog, ...)
}
\arguments{
\item{demog}{DemogParams object with a \code{StageStructure}}

\item{K_vector}{K-values (corresponding to 1/b) to run the matrix model for}

\item{plot}{plot the equilibrium population? (default is \code{TRUE})}

\item{stages_out}{which stages to plot? (defaults to all)}

\item{juv.stage}{use explicit juvenile (zeroth) stage? (default is \code{TRUE})}

\item{t_rec}{time steps to record (defaults to \eqn{1}); if \code{t_rec}\eqn{>1}, the mean over all time steps is returned}

\item{N_0}{initial condition, i.e. population at time zero; must include stage zero regardless of the value of \code{juv.stage}}

\item{t_max}{allowed number of time steps to reach equilibrium (default is \eqn{1000})}

\item{delta}{tolerance to define equilibrium (default is \eqn{.1}); the utilised measure is euclidian distance of current to previous time steps}

\item{diagnostics}{in addition to recorded population vectors, returns the number of steps taken as well as the transition matrix and the value if delta at the last step (default is \code{FALSE})}
}
\description{
Uses the Rangeshifter Demography module to create the corresponding matrix model and runs it until equilibrium is reached.
This corresponds to a population in a single cell without dispersal.
Since the matrix model representation is used, some options (e.g. maximum age) of the \code{\link[RangeshiftR]{Demography}} module can not be taken into account.
}
\details{
Rangeshifter requires an additional juvenile stage to be added to the common transition matrix as stage 0 (in order
to allow for juvenile dispersal). For the simulation with \code{RunMatrixModel()}, this stage can be kept (\code{juv.stage=TRUE})
or remove to yield the corresponding Lefkovitch matrix (\code{juv.stage=FALSE}).\cr
The default initial condition \code{N_0} is a population at its respective carrying capacity with unpopulated juvenile stage and
all higer stages equally populated.\cr
}
