% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ManagementParams.R
\docType{class}
\name{Harvesting}
\alias{Harvesting}
\title{Harvesting Parameters}
\usage{
Harvesting(years = 1,
                     HarvestMat = 0,
                     harvesting_success = 1.0
                      )
}
\arguments{
\item{years}{Vector of years in which harvesting events should take place}

\item{HarvestMat}{Matrix of harvesting events. Each row represents a harvesting event. Columns represent:

 - the year of the event,\cr
 - the location (\code{Patch_ID} in case of patch-based models or \code{X} and \code{Y} location in case of cell-based models),\cr
 - the number of individuals which are tried to be harvested,\cr
 - minimal age of each individual,\cr
 - maximal age of the individual,\cr
 - stage of the individual,\cr
 - sex of the individual\cr}

\item{harvesting_success}{Harvesting success rate}
}
\value{
a parameter object of class "ManagementParams"
}
\description{
Harvesting is a management strategy to remove individuals from a population.
Harvesting can be applied to control population size, manage invasive species, for economic purposes
but can also represent illegal poaching events.
}
\details{
\strong{General information}

This function harvests a specified number of individuals with given characteristics from a given site
in a specified year representing a harvesting event.
In other words, a harvesting event is defined by a unique combination of the year, site, and the number and characteristics of selected individuals.

You must set at least one harvesting event, but there is no upper limit. Multiple harvesting events can occur
per year (differing in either the characteristics of the individuals selected or
the sites), or across different years.

Individuals can be selected according to their age, stage and sex, depending on the type of demographic model applied.
Only individuals which are currently not in the transfer phase are eligible for harvesting.
After being selected for harvest, individuals are extracted and set to status 11.

Additionally, you can define a constant harvesting success rate to determine the likelihood of successfully extracting a selected individual.

\strong{Setting harvesting parameters}

You need to create a Harvesting Matrix \code{HarvestMat} which hold the information for each harvesting event in each row.

In the columns of the \code{HarvestMat} the year, site as well as the number of individuals and the characteristics are
defined in the following order:

- \code{year} of the translocation event \cr
- \code{Patch_ID} or \code{X} and \code{Y} location site from which individuals are harvested. \cr
- \code{nb_harvest} how many individuals of the given set of characteristics are tried to be harvested \cr
- \code{min_age} minimal age of the individual in the interval of 0-MaxAge. Set to -9 to ignore. \cr
- \code{max_age} maximal age of the individual in the interval of \code{min_age}-MaxAge. Set to -9 to ignore.\cr
- \code{stage} of the individual. Set to -9 to ignore. \cr
- \code{sex} of the individual: Only for sexual models, otherwise set to -9 to ignore  \cr

\code{min_age}, \code{max_age} and \code{stage} can only be defined for stage structured models.
For non stage structured models, or to ignore the characteristic, set the value to -9.

\code{sex} can only be defined for sexual models. For non sexual models, or to ignore the characteristic, set the value to -9.

To avoid unused harvest events, you should set the range of allowed characteristics as broad as possible.

Each row of the harvesting matrix \code{HarvestMat} should hold a unique combination of year, site,
the number of individuals to harvest as well as characteristics of these individuals.
You may add more than one harvest event per year, i.e. there may be multiple sites for each year of a harvest
event or multiple individual characteristics for a certain site.

In each \code{year} of a translocation event, individuals matching the given criteria in the site are collected and a given number of individuals \code{nb_harvest}
are sampled from this subset. The success of harvesting one of these sampled individuals in the site is defined by the \code{harvesting_success}.
Successfully caught individuals are then extracted from the simulation and the individual will be assigned the status 11 (harvested).

The site is required to be habitat patches or cells of the landscape. Otherwise the harvesting event will be skipped.
}
\references{
\insertAllCited{}
}
\author{
Jette Wolff
}
