% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_GeneticsParams.R
\docType{class}
\name{NeutralTraits}
\alias{NeutralTraits}
\title{Set genetic traits structure for neutral traits}
\arguments{
\item{Positions}{Loci positions coding for trait within genome. Must be in the
range 0-(GenomeSize-1), specified in Genetics file. Positions can overlap across
traits - there will be no pleiotropy, but this will influence genetic linkage.,}

\item{NbOfPositions}{Only specify if above is set to ‘random’, else must be blank (NULL)}

\item{InitialDistribution}{Distribution from which to draw initial allele values from.
If \code{uniform}. Initialise with random characters between 0 – max. Can be left blank (NULL)
and assume every individual is the ‘wildtype’ (max value provided in  \code{InitialParameters} ) and new mutations alter that.}

\item{InitialParameters}{Maximal value for the uniform distribution.}

\item{MutationDistribution}{Distribution for mutations to draw from. Can be either 'KAM' or 'SSM'. KAM (k-alleles model) is randomly
draw a value between 0 and max (see MutationParameters).
SSM (single-step mutation) is to move in a stepwise manner, A to B, B to C.}

\item{MutationParameters}{Parameters for the above distribution: maximal value for KAM or SSM (cannot exceed 255)}

\item{MutationRate}{Mutation rate applicable to this type of loci. Must be between 0.0 and 1.0}

\item{OutputValues}{If OutputGeneValues in \code{\link[RangeShiftR]{Genetics}} is
enabled, should allele values for this gene be written to output? Ignored if OutputGeneValues is set to \code{FALSE}.}
}
\value{
a parameter object of class "NeutralTraitsParams"
}
\description{
Set genetic traits structure for neutral traits
}
\author{
Jette Reeg
}
