% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RangeShiftR.R
\docType{package}
\name{RangeShiftR-package}
\alias{RangeShiftR}
\alias{RangeShiftR-package}
\title{RangeShiftR: An R interface to the RangeShifter simulation software}
\description{
RangeShiftR: an R package for individual-based simulation of spatial eco-evolutionary dynamics and species' responses to environmental changes
}
\details{
Bocedi, Greta, et al. "RangeShifter: a platform for modelling spatial eco‐evolutionary dynamics and species' responses to environmental changes." Methods in Ecology and Evolution 5.4 (2014): 388-396.
}
\section{Warning}{

You must not call this function unless ...

\subsection{Exceptions}{
   Apart from the following special cases...
}
}

\references{
\url{https://en.wikipedia.org/wiki/List_of_Crayola_crayon_colors}
}
\author{
Anne-Kathleen Malchow, \email{malchowa@geo.hu-berlin.de}
}
\keyword{internal}
