% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{RunRS}
\alias{RunRS}
\title{Run the Simulation}
\usage{
RunRS(RSparams, dirpath = getwd())
}
\arguments{
\item{RSparams}{A parameter master object (class 'RSparams'), contains all parameters to specify the simulation settings.}

\item{dirpath}{File path to RS working directory; must contain the folders 'Inputs', 'Outputs', 'Output_Maps'.\cr
If NULL, the current \code{R} working directory will be used.}
}
\value{
returns an error code or, if \code{RSparams@simul@ReturnPopRaster = TRUE}, a data frame with population data output
}
\description{
Run the Simulation
}
