% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_GeneticsParams.R
\docType{class}
\name{SMSTraits}
\alias{SMSTraits}
\title{Set genetic traits structure for SMS traits}
\value{
a parameter object of class "SMSTraitsParams"
}
\description{
Can I have a list of different SMStraits() depending on the dispersal parameter? (E0. Ealpa, Ebeta,)
}
\author{
Jette Reeg
}
