% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ManagementParams.R
\docType{class}
\name{Translocation}
\alias{Translocation}
\title{Translocation Parameters}
\usage{
Translocation(years = 1,
                     TransLocMat = 0,
                     catching_rate = 0.5
                      )
}
\arguments{
\item{years}{Vector of years in which translocation events should take place}

\item{TransLocMat}{Matrix of translocation events. Each row represents a translocation event. Columns represent:

 - the year of the event,\cr
 - the source location (\code{Patch_ID} in case of patch-based models or \code{X} and \code{Y} location in case of cell-based models),\cr
 - the destination location (\code{Patch_ID} in case of patch-based models or \code{X} and \code{Y} location in case of cell-based models),\cr
 - the number of individuals which are tried to be caught,\cr
 - minimal age of each individual,\cr
 - maximal age of the individual,\cr
 - stage of the individual,\cr
 - sex of the individual\cr}

\item{catching_rate}{Catching success rate}
}
\value{
a parameter object of class "ManagementParams"
}
\description{
Translocation is a conservation management strategy to reintroduce species in areas where they have been extirpated,
stabilize metapopulations or enhance (genetic) connectivity between fragmented habitats.
Translocations aim to increase the survival probabilities of species, boost genetic diversity,
and restore ecological balance. Success depends on careful selection of suitable source and destination sites, as well as
appropriate selection of individuals to be relocated.
}
\details{
\strong{General information}

This function translocates a specified number of individuals with given characteristics from a source site
to a destination site in a specified year representing a translocation event.
In other words, a translocation event is defined by a unique combination of the year, source and
destination site, and the number and characteristics of selected individuals.

You must set at least one translocation event, but there is no upper limit. Multiple translocation events can occur
per year (differing in either the characteristics of the individuals selected or
the source and destination sites), or across different years.

Individuals can be selected according to their age, stage and sex, depending on the type of demographic model applied.
Only individuals which are currently not in the transfer phase are eligible for translocation.
After being translocated, individuals are assumed to not disperse further.

Additionally, you can define a constant catching success rate to determine the likelihood of successfully capturing a selected individual.

\strong{Setting translocation parameters}

You need to create a Translocation Matrix \code{TransLocMat} which hold the information for each translocation event in each row.

In the columns of the \code{TransLocMat} the year, source and destination site as well as the number of individuals and the characteristics are
defined in the following order:

- \code{year} of the translocation event \cr
- \code{Patch_ID} or \code{X} and \code{Y} location of the source site from which individuals are caught. \cr
- \code{Patch_ID} or \code{X} and \code{Y} location of the destintation site to which individuals are relocated. \cr
- \code{nb_catch} how many individuals of the given set of characteristics are tried to be caught \cr
- \code{min_age} minimal age of the individual in the interval of 0-MaxAge. Set to -9 to ignore. \cr
- \code{max_age} maximal age of the individual in the interval of \code{min_age}-MaxAge. Set to -9 to ignore.\cr
- \code{stage} of the individual. Set to -9 to ignore. \cr
- \code{sex} of the individual: Only for sexual models, otherwise set to -9 to ignore  \cr

\code{min_age}, \code{max_age} and \code{stage} can only be defined for stage structured models.
For non stage structured models, or to ignore the characteristic, set the value to -9.

\code{sex} can only be defined for sexual models. For non sexual models, or to ignore the characteristic, set the value to -9.

To avoid unsuccessful translocation events, you should set the range of allowed characteristics as broad as possible.

Each row of the translocation matrix \code{TransLocMat} should hold a unique combination of a source and a destination location,
the number of individuals to catch as well as characteristics of these individuals.
You may add more than one translocation event per year, i.e. there may be multiple source and destination sites for each year of a translocation
event or multiple individual characteristics for a certain pair of source and destination site.

In each \code{year} of a translocation event, individuals matching the given criteria in the source site are collected and a given number of individuals \code{nb_catch}
are sampled from this subset. The success of catching one of these sampled individuals in the source location is defined by the \code{catching_rate}.
Successfully caught individuals are then translocated to the destination site and the individual will be assigned the status 10 (translocated) and will not further disperse.

The source as well as the destination sites are required to be habitat patches or cells of the landscape. Otherwise the translocation event will be skipped.
}
\references{
\insertAllCited{}
}
\author{
Jette Wolff
}
