% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_handling.R
\name{createParameterTables}
\alias{createParameterTables}
\title{## ---- ODD documentation -----}
\usage{
createParameterTables(filename, s, type, title)
}
\arguments{
\item{filename}{Name of the R markdown file and document to be created, e.g. 'Parameter_table.rmd'}

\item{s}{RangeShiftR parameter object}

\item{type}{file type of the rendering process output. Can be either "pdf_document", "word_document" or "md_document" We recommend using "word_document" to refine the layout.}

\item{title}{Title of the study if possible including authors, e.g. "'Study title ABC' by author ABC}
}
\description{
#' Creating ODD template file for a specific RangeShiftR parameter master object
#'
#' This function creates an ODD template file for a specific RangeShiftR parameter master object \code{s}.
#' It only creates a new file, if the \code{filename} doesn't exist in the folder.
#' If the \code{filename} already exists, it only renders the document to the given type.
#'
#' @usage createODD(filename, s, type, title)
#'
#' @param filename Name of the R markdown file and document to be created, must have ending ".Rmd", e.g. 'ODD_protocol.Rmd'
#' @param s RangeShiftR parameter object
#' @param type file type of the rendering process output. Can currently only be "pdf_document"
#' @param title Title of the study if possible including authors, e.g. "'Study title ABC' by author ABC
#' @export
setGeneric("createODD", function(filename, s, type, title,...) standardGeneric("createODD") )
}
\details{
setMethod("createODD", c(filename = "character", s="RSparams", type="character", title="character"), function(filename="ODD_protocol_template.Rmd", s, type="pdf_document", title = "'Study ABC' by author ABC"){
    if(!file.exists(filename)) {
        unlink(c("RSflowchart_big.pdf", "RSflowchart_detail.pdf", "RSflowchart_big.svg", "RSflowchart_detail.svg", "style-template.docx", "RS_ODD.json", "ecography.csl", "RS_ODD.bib"))
        rmarkdown::draft(filename, template = "odd_protocol", package = "RangeShiftR", edit = FALSE)
    }
    if (type=="pdf_document") format <- "pdf"
    if (type=="md_document") format <- "md"
    if (type=="rtf_document") format <- "word"
    if (type=="word_document") format <- "word"

    subtitle <- title

    rmarkdown::render(input = filename, output_format = type, params=list(format = format, subtitle = title))
})
Creating parameter table file for a specific RangeShiftR parameter master object

This function creates a file with tables of the parameter set for a specific RangeShiftR parameter master object \code{s}.
It only creates a new file, if the \code{filename} doesn't exist in the folder.
If the \code{filename} already exists, it only renders the document to either pdf, word or md.
}
