% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_GeneticsParams.R
\docType{class}
\name{EmigrationTraits}
\alias{EmigrationTraits}
\title{Set genetic traits structure for emigration traits}
\usage{
EmigrationTraits(Positions = list("random"), NbOfPositions = 10,
ExpressionType = "additive",
InitialDistribution = "normal", InitialParameters = matrix(c(0.5,0.1), nrow=1),
IsInherited = FALSE,
MutationDistribution = "normal", MutationParameters = matrix(c(0.5,0.2), nrow=1),
MutationRate = 0.0001, OutputValues = FALSE)
}
\arguments{
\item{Positions}{Loci positions coding for the trait within genome. Should be provided as a list. Entries can either be a string (\code{"random"}) and/or vectors of integers.
The length must be equal to the number of required emigration traits (see above) and the sequence must match the sequence of the emigration traits listed above.}

\item{NbOfPositions}{Only specify when the \code{Positions} of the emigration trait is set to \code{"random"}, else must be blank (\code{NULL}).
The length must be equal to the number of required emigration traits (see above) and the sequence must match the sequence of the emigration traits listed above.}

\item{ExpressionType}{Type of expression for the emigration trait. Can be either \code{additive} or \code{average}.
The length must be equal to the number of required emigration traits (see above) and the sequence must match the sequence of the emigration traits listed above.}

\item{InitialDistribution}{Distribution of the initial values. Can be \code{uniform} or \code{normal}. Should be provided as a vector of strings.
The length must be equal to the number of required emigration traits (see above) and the sequence must match the sequence of the emigration traits listed above.}

\item{InitialParameters}{Parameters for the initial distribution: You must provide two colums min and max  for \code{uniform} distribution and mean and sd for \code{normal} distribution.
Each row in the matrix corresponds to an emigration trait. The number of rows must be equal to the number of required emigration traits (see above) and the sequence must match the sequence of the emigration traits listed above.}

\item{IsInherited}{Should the emigration trait be inherited? Can be either \code{TRUE} or \code{FALSE}.
The length must be equal to the number of required emigration traits (see above) and the sequence must match the sequence of the emigration traits listed above.}

\item{MutationDistribution}{Distribution for mutations to draw from. Can be \code{uniform} or \code{normal}.
The length must be equal to the number of required emigration traits (see above) and the sequence must match the sequence of the emigration traits listed above.}

\item{MutationParameters}{Parameters for the mutation distribution: You must provide two colums: min and max for \code{uniform} distribution and mean and sd for \code{normal} distribution.
Each row in the matrix corresponds to an emigration trait. The number of rows must be equal to the number of required emigration traits (see above) and the sequence must match the sequence of the emigration traits listed above.}

\item{MutationRate}{Mutation rate applicable to this type of loci. Must be between \code{0.0} and \code{1.0}.
The length must be equal to the number of required emigration traits (see above) and the sequence must match the sequence of the emigration traits listed above.}

\item{OutputValues}{If OutputGeneValues in GeneticsFile is enabled, should allele values for this gene be written to output? Ignored if OutputGeneValues is set to \code{FALSE}.
The length must be equal to the number of required emigration traits (see above) and the sequence must match the sequence of the emigration traits listed above.}
}
\value{
a parameter object of class "EmigrationTraitsParams"
}
\description{
Depending of the settings in \code{\link[RangeShiftR]{Emigration}}, up to three emigration traits evolve:

- The probability of emigration \ifelse{html}{\out{D<sub>0</sub>}}{\eqn{D_0}} / \eqn{d} \cr
- The slope of the density dependent emigration function (if \code{DensDep=TRUE}) \ifelse{html}{\out{&alpha;}}{\eqn{α}} \cr
- The density threshold of the density dependent emigration function (if \code{DensDep=TRUE}) \ifelse{html}{\out{&beta;}}{\eqn{β}} \cr

If emigration is sex dependent, \code{SexDep=TRUE}, you must provide details for both sexes.

This results in following number of emigration traits that need to be specifed:

- 1 entry if emigration is neither density dependent nor sex dependent (\code{DensDep=FALSE} and \code{SexDep=FALSE}): emgiration probability \eqn{d} \cr
- 2 entries if emigration is sex dependent (\code{SexDep=TRUE}): emigration probability of females \eqn{d(f)} and males \eqn{d(m)} \cr
- 3 entries if emigration probability is only density dependent \code{DensDep=TRUE}:  \ifelse{html}{\out{D<sub>0</sub>}}{\eqn{D_0}}, \ifelse{html}{\out{&alpha;}}{\eqn{α}} and \ifelse{html}{\out{&beta;}}{\eqn{β}} \cr
- 6 entries if emigration probability is both density dependent (\code{DensDep=TRUE}) and sex dependent (\code{SexDep=TRUE}): \ifelse{html}{\out{D<sub>0</sub>(f), D<sub>0</sub>(m)}}{\eqn{D_0(f),D_0(m)}}, \ifelse{html}{\out{&alpha;(f), &alpha;(m)}}{\eqn{α(f), α(m)}} and \ifelse{html}{\out{&beta;(f), &beta;(m)}}{\eqn{β(f), β(m)}} \cr

The entries of the trait parameters must be provided in the same order as the emigration traits are listed above. If parameters expect a matrix, the rows must match the order of kernel traits listed above.
}
\details{
Traits set to evolve cannot simultaneously be stage-dependent.

The alleles of each trait can be expressed according to an additive model (allele values across all loci are summed) or be averaged.

Mutations are additive and can be sampled in either a \code{normal} or \code{uniform} distribution.

Initial allele values are sampled in a \code{normal} or \code{uniform} distribution.

Emigration traits can also be **not** inherited, that is, allele values are resampled from the initial distribution for every new individual.

Dominance values are not applicable for emigration traits.
}
\author{
Jette Reeg
}
