% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_GeneticsParams.R
\docType{class}
\name{Genetics}
\alias{Genetics}
\title{Set Genetics parameters}
\usage{
Genetics(GenomeSize = 10,
         ChromosomeEnds = 1, RecombinationRate = 0.0,
         OutputGeneValues = FALSE, OutputNeutralStatistics = FALSE,
         OutputFstatsWeirCockerham = FALSE, OutputFstatsWeirHill = FALSE,
         OutputStartGenetics = NULL, OutputInterval = NULL,
         PatchList = NULL, NbrPatchToSample = NULL,
         nIndividualsToSample = NULL, Stages = NULL, Traits = Traits()
         )
}
\arguments{
\item{GenomeSize}{Maximum size of genome (number of loci)}

\item{ChromosomeEnds}{Where the genome is split into chromosomes, if empty
assumed one chromosome is equal to GenomeSize. These areas recombine with a
probability of 0.5.}

\item{RecombinationRate}{Recombination rate (through chromosomal crossover)
across the whole genome (in addition to the chromosomeEnds above).}

\item{OutputGeneValues}{Output the values of all alleles for all genes of all
sampled individuals. Does not output the resulting trait values: mean and SD
of dispersal and genetic fitness traits are output in the TraitsXPatch,
TraitsXCell and/or TraitsXrow output files. Enables the geneValues output
files.}

\item{OutputFstatsWeirCockerham}{Calculate F-statistics (including global and
per-locus estimates) according to Weir & Cockerham (1984)'s method-of-moments
approach. Enables the neutralGenetics and perLocusNeutralGenetics output files.}

\item{OutputFstatsWeirHill}{Calculate F-statistics calculated according to the
estimators of Weir & Hill (2002), including global estimates corrected for
unequal sample sizes, population- (i.e. patch-) specific estimates, and pairwise
estimates. Enables the neutralGenetics and pairwisePatchNeutralGenetics output files.}

\item{OutputStartGenetics}{Which year should RangeShifter start to produce the output files listed above?}

\item{OutputInterval}{How frequently to output genetic output, including gene values and neutral statistics.}

\item{PatchList}{Which patches are to be sampled for output.  Patches can be
specified according to their patch number, as per the patch layer in a patch-based
model. Or sampled randomly or all patches can be chosen. In a cell-based landscape
random is the only option with number of patches (=cells) specified.}

\item{NbrPatchToSample}{If PatchList=random or random_occupied then this specifies
the number of patches to sample randomly. Random: The chosen sample patches remain
the same throughout the simulation, i.e. do not vary between years or replicates
unless artificially generated landscape that is generated afresh between replicates.
Random_occupied: patches are re-sampled every generation among all patches containing at least 1 individual.}

\item{nIndividualsToSample}{The number of individuals to sample in a patch. If nInds < nIndividualsToSample then sampled individuals = nInds}

\item{Stages}{The age stages to sample from.}

\item{Traits}{The genetic traits to be modelled.}
}
\value{
a parameter object of class "GeneticsParams"
}
\description{
Set genetics parameters\cr

Controls heritability and evolution of traits (if inter-individual variability is enabled (\code{IndVar=TRUE}) for at least one (dispersal) trait).
Provides control over the genome size, the number of loci, the recombination rate (if the species is diploid) and a
flexible mapping of traits to chromosomes, allowing linkage, pleiotropy and neutral alleles to be incorporated. It is also possible to model
neutral alleles when no adaptive traits are present.
}
\details{
TBD
}
\author{
Jette Reeg
}
