% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_GeneticsParams.R
\docType{class}
\name{KernelTraits}
\alias{KernelTraits}
\title{Set genetic traits structure for kernel traits}
\value{
a parameter object of class "KernelTraitsParams"
}
\description{
Can I have a list of different kerneltraits() depending on the dispersal parameter? (E0. Ealpa, Ebeta,)
}
\author{
Jette Reeg
}
