% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ManagementParams.R
\docType{class}
\name{Management}
\alias{Management}
\title{Management Parameters}
\usage{
Management(Translocation = Translocation())
}
\arguments{
\item{Translocation}{Set translocation events. See \code{\link{Translocation}} for more details.}
}
\value{
a parameter object of class "ManagementParams"
}
\description{
With this function you can set all management strategies you wish to apply to your simulations.
This function includes for now only the translocation of individuals (see details below).
}
\details{
\strong{Translocation}

Translocation is a conservation management strategy to reintroduce species in areas where they have been extirpated,
stabilize metapopulations or enhance (genetic) connectivity between fragmented habitats. Translocations are applied
to increase survival probabilities of species, increase genetic diversity or restore ecological balance.The selection of
suitable source and destination sites as well as selection of suitable individuals is crucial for the success of the strategy.
}
\references{
\insertAllCited{}
}
\author{
Jette Reeg
}
