% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ManagementParams.R
\docType{class}
\name{Management}
\alias{Management}
\title{Set Management Parameters}
\usage{
Management(Translocation = Translocation())
}
\arguments{
\item{Translocation}{Individuals are selected in one location and then transfered to a new location. You can define the year(s) of the translocation event(s),
the source and target location, the number of individuals to be translocated and characteristics of individuals to be sampled from. The catching rate defines the success catching an individual.
See \code{\link{Translocation}} for more details.}
}
\value{
a parameter object of class "ManagementParams"
}
\description{
Set all management parameters you wish to apply on your simulations. This includes for now only the translocation of individuals.
}
\details{
TODO: Add general information on management options and translocations
}
\references{
\insertAllCited{}
}
\author{
Jette Reeg
}
