% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_GeneticsParams.R
\docType{class}
\name{NeutralTraits}
\alias{NeutralTraits}
\title{Set up structure of neutral traits}
\arguments{
\item{Positions}{Loci positions coding for trait within genome. Must be in the
range 0-(\code{GenomeSize}-1), specified in \code{\link[RangeShiftR]{Genetics}}. Positions can overlap across
traits - there will be no pleiotropy, but this will influence genetic linkage.,}

\item{NbOfPositions}{Only specify if \code{Positions} is set to \code{"random"}, else must be blank (\code{NULL})}

\item{InitialPositions}{Specify which positions should be initialised. Must be "all", "random" or a vector of integers. If enabled, the remaining initial value parameters will only be applied to the selected positions,
and all other positions will be set to 0. Must be in the range  of \code{Positions}.}

\item{InitialNbOfPositions}{Only specify if \code{InitialPositions} is set to \code{"random"}, else must be blank (\code{NULL}). Cannot be greater than \code{NbOfPositions}}

\item{InitialDistribution}{Distribution from which to draw initial allele values from.
If \code{uniform}. Initialise with random characters between 0 – \code{max}. Note that possible values start at 0, so \code{max=0}
specifies a monomorphic initial population.}

\item{InitialParameters}{Maximal value for the uniform distribution.}

\item{MutationDistribution}{Distribution for mutations to draw from. Can be either \code{"KAM"} or \code{"SSM"}. \code{KAM} (k-alleles model) is randomly
draw a value between 0 and \code{max} (see \code{MutationParameters}).
\code{SSM} (single-step mutation) is to move in a stepwise manner, A to B, B to C.}

\item{MutationParameters}{Parameters for the above distribution: maximal value for \code{KAM} or \code{SSM} (cannot exceed 255)}

\item{MutationRate}{Mutation rate applicable to this type of loci. Must be between 0.0 and 1.0}

\item{OutputValues}{If OutputGeneValues in \code{\link[RangeShiftR]{Genetics}} is
enabled, should allele values for this gene be written to output? Ignored if OutputGeneValues is set to \code{FALSE}.}
}
\value{
a parameter object of class "NeutralTraitsParams"
}
\description{
A method to specify neutral traits in the genetic module.
}
\details{
#' @usage NeutralTraits(Positions = "random", NbOfPositions = 10,
Positions = "random", # "random" or list of integer values
NbOfPositions = 10, # numeric, only of positions random
InitialPositions = "all", # "random" or list of integer values
InitialNbOfPositions = NULL, # numeric, only of positions random
InitialDistribution = NULL, # uniform (neutral + dispersal), normal (dispersal)
InitialParameters = 2, # neutral traits: only max value; dispersal: two values: either min/max oder mean+sd, not applicable for genetic load
MutationDistribution = "KAM", # neutral: "KAM" or "SSM", genetic load: "gamma", "uniform", "normal", "negExp", dispersal: uniform or normal
MutationParameters = 2, # single value or 2 values
MutationRate = 0.0, # numeric
OutputValues = FALSE)



Neutral trait does not have any phenotypic effect during the simulation. It is used to compute F-statistics and other measures of neutral variation.
For neutral trait you must specify:

-	The number and positions of genes controlling the trait: \code{Positions} and \code{NbOfPositions} \cr
-	A distribution to sample initial values from: \code{InitialDistribution} and \code{InitialParameters} \cr
-	A mutation rate for all genes controlling the trait. \code{MutationRate} \cr
-	A distribution to sample mutations from. \code{MutationDistribution} and \code{MutationParameters} \cr

The user specifies the number of possible alleles for neutral loci (up to 256), via the maximum parameter of the mutation distribution.
Initial values are either identical for all sites (equal to the max value) or sampled in a uniform distribution (between 0 and the maximum value).
Mutations are either sampled in a uniform distribution between 0 and the max parameter (k-allele model, KAM, \insertCite{peng2012}{RangeShiftR}) or
added as increments (random -1 or +1 changes) of the previous value (symmetric stepwise model, SSM, \insertCite{peng2012}{RangeShiftR}).

Dominance values and inheritance are not applicable for neutral traits.
}
\references{
\insertAllCited{}
}
\author{
Jette Reeg
}
