% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RangeshiftR.R
\docType{package}
\name{RangeshiftR-package}
\alias{RangeshiftR}
\alias{RangeshiftR-package}
\title{RangeshiftR: An R interface to the RangeShifter simulation software}
\description{
RangeShifter: Modelling eco-evolutionary dispersal dynamics & Investigating species’ responses to environmental changes.
	This package contains the simulation software (as C++ code) and R-level entry functions to set parameters and start simulations, as well as some output handling functions.
}
\details{
The only function you're likely to need from roxygen2 is [roxygenize()].
Otherwise refer to the vignettes to see how to format the documentation.

Bocedi, Greta, et al. "RangeShifter: a platform for modelling spatial eco‐evolutionary dynamics and species' responses to environmental changes." Methods in Ecology and Evolution 5.4 (2014): 388-396.
}
\section{Warning}{

You must not call this function unless ...

\subsection{Exceptions}{
   Apart from the following special cases...
}
}

\references{
\url{https://en.wikipedia.org/wiki/List_of_Crayola_crayon_colors}
}
\author{
Anne-Kathleen Malchow, \email{malchowa@geo.hu-berlin.de}
}
\keyword{internal}
