% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_GeneticsParams.R
\docType{class}
\name{SettlementTraits}
\alias{SettlementTraits}
\title{Set genetic traits structure for neutral traits}
\value{
a parameter object of class "SettlementTraitsParams"
}
\description{
Can I have a list of different dispersaltraits() depending on the dispersal parameter? (E0. Ealpa, Ebeta,)
}
\author{
Jette Reeg
}
