% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_GeneticsParams.R
\docType{class}
\name{SMSTraits}
\alias{SMSTraits}
\title{Set genetic traits structure for SMS traits}
\usage{
SMSTraits(Positions = list("random"), NbOfPositions = c(10),
ExpressionType = rep("additive",1),
InitialDistribution = rep("normal",1), InitialParameters = matrix(c(rep(0.5,1),(rep(0.1,1), nrow=1),
IsInherited = rep(FALSE,1),
MutationDistribution = rep("normal",1), MutationParameters = matrix(c(rep(0.5,1),(rep(0.2,1), nrow=1),
MutationRate = rep(0.0001,1), OutputValues = rep(FALSE,1))
}
\arguments{
\item{Positions}{Loci positions coding for the trait within genome. Should be provided as a list. Entries can either be a string (\code{"random"}) and/or vectors of integers.
The length must be 1 (\code{GoalType=0}) or 4 (\code{GoalType=2}) for the required SMS traits \code{DP} and, if \code{GoalType=2} \code{GoalBias}, , \code{AlphaDB} and , \code{BetaDB}.}

\item{NbOfPositions}{Only specify when the \code{Positions} of the SMS trait is set to ‘random’, else must be blank (\code{NULL}).
The length must be 1 (\code{GoalType=0}) or 4 (\code{GoalType=2}) for the required SMS traits \code{DP} and, if \code{GoalType=2} \code{GoalBias}, , \code{AlphaDB} and , \code{BetaDB}.}

\item{ExpressionType}{Type of expression for the emigration trait. Can be either \code{additive} or \code{average}.
The length must be 1 (\code{GoalType=0}) or 4 (\code{GoalType=2}) for the required SMS traits \code{DP} and, if \code{GoalType=2} \code{GoalBias}, , \code{AlphaDB} and , \code{BetaDB}.}

\item{InitialDistribution}{Distribution of the initial values. Can be \code{uniform} or \code{normal}. Should be provided as a vector of strings.
The length must be 1 (\code{GoalType=0}) or 4 (\code{GoalType=2}) for the required SMS traits \code{DP} and, if \code{GoalType=2} \code{GoalBias}, , \code{AlphaDB} and , \code{BetaDB}.}

\item{InitialParameters}{Parameters for the initial distribution: You must provide two colums min and max  for \code{uniform} distribution and mean and sd for \code{normal} distribution.
Each row in the matrix corresponds to an SMS trait.
The number of rows must be 1 (\code{GoalType=0}) or 4 (\code{GoalType=2}) for the required SMS traits \code{DP} and, if \code{GoalType=2} \code{GoalBias}, , \code{AlphaDB} and , \code{BetaDB}.}

\item{IsInherited}{Should the emigration trait be inherited? Can be either \code{TRUE} or \code{FALSE}.
The length must be 1 (\code{GoalType=0}) or 4 (\code{GoalType=2}) for the required SMS traits \code{DP} and, if \code{GoalType=2} \code{GoalBias}, , \code{AlphaDB} and , \code{BetaDB}.}

\item{MutationDistribution}{Distribution for mutations to draw from. Can be \code{uniform} or \code{normal}.
The length must be 1 (\code{GoalType=0}) or 4 (\code{GoalType=2}) for the required SMS traits \code{DP} and, if \code{GoalType=2} \code{GoalBias}, , \code{AlphaDB} and , \code{BetaDB}.}

\item{MutationParameters}{Parameters for the mutation distribution: You must provide two colums: min and max for \code{uniform} distribution and mean and sd for \code{normal} distribution.
Each row in the matrix corresponds to an SMS trait.
The number of rows must be 1 (\code{GoalType=0}) or 4 (\code{GoalType=2}) for the required SMS traits \code{DP} and, if \code{GoalType=2} \code{GoalBias}, , \code{AlphaDB} and , \code{BetaDB}.}

\item{MutationRate}{Mutation rate applicable to this type of loci. Must be between 0.0 and 1.0.
The length must be 1 (\code{GoalType=0}) or 4 (\code{GoalType=2}) for the required SMS traits \code{DP} and, if \code{GoalType=2} \code{GoalBias}, , \code{AlphaDB} and , \code{BetaDB}.}

\item{OutputValues}{If OutputGeneValues in GeneticsFile is enabled, should allele values for this gene be written to output? Ignored if OutputGeneValues is set to \code{FALSE}.
The length must be 1 (\code{GoalType=0}) or 4 (\code{GoalType=2}) for the required SMS traits \code{DP} and, if \code{GoalType=2} \code{GoalBias}, , \code{AlphaDB} and , \code{BetaDB}.}
}
\value{
a parameter object of class "SMSTraitsParams"
}
\description{
Depending on the settings of \code{\link[RangeShiftR]{SMS}}, the following traits can be evolvable:

- The directional persistence of the individual \code{DP} \cr
- The \code{GoalBias}, that is the tendency for an individual to move away from its natal patch \cr
- The slope of the distance dependent decay of the goal bias \code{AlphaDB} \cr
- The distance threshold of the distance dependent devay of the goal bias \code{BetaDB} \cr

If dispersal kernels are sex dependent, \code{SexDep=TRUE}, you must provide details for both sexes.

This results in the following number of SMS traits that need to be specified:

- 1 if \code{GoalType = 0}: directional persistence \code{DP} \cr
- 4 if \code{GoalType = 2}: \code{DP}, \code{GoalBias}, \code{AlphaDB}, \code{BetaDB} \cr

The entries of the trait parameters must be provided in the same order as the SMS traits are listed above. If parameters expect a matrix, the rows must match the order of SMS traits listed above.
}
\details{
Traits set to evolve cannot simultaneously be stage-dependent.

The alleles of each trait can be expressed according to an additive model (allele values across all loci are summed) or be averaged.

Mutations are additive and can be sampled in either a \code{normal} or \code{uniform} distribution.

Initial allele values are sampled in a \code{normal} or \code{uniform} distribution.

SMS traits can also be **not** inherited, that is, allele values are resampled from the initial distribution for every new individual.

Dominance values are not applicable for SMS traits.
}
\author{
Jette Reeg
}
