% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_GeneticsParams.R
\docType{class}
\name{Traits}
\alias{Traits}
\title{Set genetic traits structure}
\value{
a parameter object of class "TraitsParams"
}
\description{
Set genetic traits structure
}
\author{
Jette Reeg
}
