% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ManagementParams.R
\docType{class}
\name{Translocation}
\alias{Translocation}
\title{Set Translocation Parameters}
\usage{
Translocation(years = 1,
                     TranLocMat = 0,
                     catching_rate = 0.5
                      )
}
\arguments{
\item{years}{Vector of years in which translocation events should take place}

\item{TranLocMat}{Matrix of translocation events. Each row represents a translocation event. Columns represent:

 - the year of the event,
 - the source location (\code{Patch_ID} in case of cell-based models or \code{X} and \code{Y} location in case of cell-based models),
 - the target location,
 - the number of individuals which are tried to be catched,
 - minimal age of each individual,
 - maximal age of the individual,
 - stage of the individual,
 - sex of the individual}

\item{catching_rate}{Catching success rate}
}
\value{
a parameter object of class "ManagementParams"
}
\description{
Translocation transfers individuals from one location to another in a given year. You should set at least one year for a translocation event,
but you can also define multiple years. You can one or multiple locations (patches or cells) from which individuals should be catched and selected for translocation.
The number and characteristics of selected individuals may vary between years. Individuals are selected according to the given minimal and maximal age, stage,
sex nd dispersal status. Additionally, you can define a catching success rate to define the success of really catching an individual.
}
\details{
In each \code{year} of a translocation event, individuals are selected from a given source location and transferred to a new target location.
The success of catching an individual in the source location is defined by the \code{catching_rate}.
To select only individuals within a certain range of characteristics, a Translocation Matrix \code{TranLocMat} is defined in which the characteristics of the individuals are defined for each translocation event.
Each row of the translocation matrix \code{TranLocMat} holds a unique combination of source and target location as well as characteristics of catched individuals.
You may add more than one translocation event per year, i.e. there may be multiple source and target locations for each year of a translocation event or multiple individual characteristics for a certain pair of source and target location.

In the columns of the \code{TranLocMat} the year, source and target location as well as the number of individuals and the characteristics are defined:

- \code{year} of the translocation event
- \code{Patch_ID} or \code{X} and \code{Y} location of the source location from which individuals are catched
- \code{Patch_ID} or \code{X} and \code{Y} location of the target location to which individuals are transferred
- \code{nb_catch} how many individuals of the given set of characteristics are tried to be catched
- \code{min_age} minimal age of the individual: 0-MaxAge, Set to 0 to ignore lower boundary
- \code{max_age} maximal age of the individual: \code{min_age}-MaxAge, , Set 0 to ignore upper boundary
- \code{stage} of the individual: Only for non-stage structured models, otherwise set to 'NA'
- \code{sex} of the individual: Only for sexual models, otherwise set to 'NA'

To avoid unsuccessful translocation events, you should set the range of allowed characteristics as broad as possible.
}
\references{
\insertAllCited{}
}
\author{
Jette Reeg
}
